agents/deepagents-simple-research/pyproject.toml
[project]
name = "deepagents-simple-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.3.9",
    "dotenv>=0.9.9",
    "tavily-python>=0.7.21",
    "shared",
    "langchain>=1.2.7",
    "langchain-openai>=0.3.0",
]

[tool.uv.sources]
shared = { path = "../../shared" }
